requirements.txt
django>=5.0.0
djangorestframework>=3.15.0
ollama>=0.2.0
docling>=2.0.0
pdf2image>=1.17.0
python-dotenv>=1.0.0
google-genai>=1.0.0
pytesseract>=0.3.10
rank_bm25>=0.2.2
sentence-transformers>=2.7.0
pdfplumber>=0.10.0
